#2 Addressed Paul comments
