+++ env.set.local0
@@ -1,0 +1,21 @@
+#!/bin/bash
+
+#! Note: Hardcoded paths are for use on the Zimmerman lab cluster, Athena, which runs CentOS7. 
+#!       Other systems may require different paths.
+
+# HVHPC 20.7-21.9 (and respective openmpi versions) have been found to successfully compile.
+module load openmpi/3.1.2-pgi-20.7
+module load pgi/20.7
+
+# Cmake 3.15 and above is required.
+module load cmake
+
+# Add PGI related environment modules to PATH; current path is Athena specific. 
+export PGI_LIBS=/export/apps/CentOS7/pgi/2020/Linux_x86_64/20.7/
+
+# Add libcint to PATH for GTO evaluation; current path is Athena specific. 
+#! Note: Libcint version 5.3.0 is required for SlaterGPU.
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/home/khoi/Modules/libraries/libcint/gcc-8.2.0/lib64
+export LIBCINT_PATH=/home/khoi/Modules/libraries/libcint/gcc-8.2.0
+
+export OMP_NUM_THREADS=2
